Add a beer page that render in HTML.
This commit adds a beer page at /beer to render a browseable format of
what we have on tap. In the future, I'd like to change this so the
default view is available at / and the json stuff is only available at a
specific end point or when specifically request json.