Enable development run mode
The goal of this commit is to enable a "developer mode" for the app, that lets you edit it and have your changes show up
immediately, instead of having to rebuild the container everytime you change something.

I changed the `ENTRYPOINT` in the Dockerfile to a `RUN` directive, allowing it to be overriden by `docker run`
commandline arguments. (The docs on this are very confusing but I think this is right:
https://docs.docker.com/engine/reference/builder/#understand-how-cmd-and-entrypoint-interact)

It also adds a development mode to the app, by installing `shotgun` at image build time and adding a `make run_dev`
target in the Makefile which mounts the host filesystem in the image and runs the app under shotgun for auto reloading
whenever the file changes.